Cleaned up and greatly simplified ROSETTA, and did other cleanup and warning removal
git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@82 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ src/midend/astUtil/annotation/OperatorDescriptors.C
@@ -117,7 +117,7 @@ bool OperatorSideEffectDescriptor::read( istream& in, const OperatorDeclaration&
 { 
   param_num = 0;
   if (BaseClass::read(in, op)) {
-    for (int i = 0; i < decl.num_of_params(); ++i) {
+    for (size_t i = 0; i < decl.num_of_params(); ++i) {
       if (contain_parameter(i))
 	++param_num;
     }
